requirements.txt
flask
google-generativeai
youtube_transcript_api
